refactor(architecture): restructure project into feature-based modules
