+++ src/main/java/com/bookreviewhub/backend/common/security/filter/JwtAuthenticationFilter.java
@@ -1,13 +1,12 @@
-package com.bookreviewhub.backend.security;
+package com.bookreviewhub.backend.common.security.filter;
 
-import com.bookreviewhub.backend.service.JwtService;
+import com.bookreviewhub.backend.common.security.jwt.JwtService;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.context.SecurityContext;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
@@ -15,7 +14,6 @@ import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
 
 import java.io.IOException;
-import java.net.SecureCacheResponse;
 
 @Component
 @RequiredArgsConstructor
